+++ inc/model/users/userRoll.php
@@ -9,7 +9,7 @@ namespace fpcm\model\users;
 
 /**
  * Benutzerrolle Objekt
- * 
+ *
  * @package fpcm\model\user
  * @author Stefan Seehafer aka imagine <fanpress@nobody-knows.org>
  * @copyright (c) 2011-2024, Stefan Seehafer
@@ -95,7 +95,7 @@ implements
     {
         return (bool) $this->is_system;
     }
-        
+
     /**
      * Set roll name
      * @param string $leveltitle
@@ -115,7 +115,7 @@ implements
     {
         $this->codex = $codex;
     }
-    
+
     /**
      * Returns translates roll name
      * @return string
@@ -136,11 +136,22 @@ implements
 
         $this->removeBannedTexts();
 
-        $params = $this->events->trigger('userroll\save', [
+        $ev = $this->events->trigger($this->getEventName('save'), [
             'leveltitle' => $this->leveltitle,
             'codex' => $this->codex,
             'is_system' => $this->is_system
-        ])->getData();
+        ]);
+
+        if (!$ev->getSuccessed() || !$ev->getContinue()) {
+            trigger_error(sprintf("Event userroll\save failed. Returned success = %s, continue = %s", $ev->getSuccessed(), $ev->getContinue()));
+            return false;
+        }
+
+        $params = $ev->getData();
+        if (!is_array($params)) {
+            trigger_error(__METHOD__ . ' save params must be an array!');
+            return false;
+        }
 
         $newId = $this->dbcon->insert($this->table, $params);
 
@@ -175,18 +186,36 @@ implements
 
         $this->dbcon->transaction();
 
-        $return = parent::delete();
+        $evbn = $this->getEventName('deleteBefore');
+        $evb = $this->events->trigger($evbn, $this->id);
+
+        if (!$evb->getSuccessed() || !$evb->getContinue()) {
+            trigger_error(sprintf("Event %s failed. Returned success = %s, continue = %s", $evbn, $evb->getSuccessed(), $evb->getContinue()));
+            return false;
+        }
+
+        $this->dbcon->delete($this->table, 'id = ?', [$this->id]);
 
         $permissions = new \fpcm\model\permissions\permissions($this->getId());
 
+        $return = false;
         if ($permissions->delete()) {
-            $return = $return && true;
+            $return = true;
         }
 
         $this->dbcon->update(\fpcm\classes\database::tableAuthors, ['roll'], [-1, $this->id], 'id = ?');
         $this->dbcon->commit();
+
+        $evan = $this->getEventName('deleteAfter');
+        $eva = $this->events->trigger($evan, $this->id);
+
+        if (!$eva->getSuccessed() || !$eva->getContinue()) {
+            trigger_error(sprintf("Event %s failed. Returned success = %s, continue = %s", $evan, $eva->getSuccessed(), $eva->getContinue()));
+            return false;
+        }
+
         $this->cache->cleanup();
-        
+
         return $return;
     }
 
@@ -207,14 +236,29 @@ implements
 
         $return = false;
 
-        $params = $this->events->trigger('userroll\update', [
+        $params = [
             'leveltitle' => $this->leveltitle,
             'codex' => $this->codex,
-        ])->getData();
-        
-        $params[] = $this->id;
+            'is_system' => $this->is_system,
+            $this->id
+        ];
 
-        if ($this->dbcon->update($this->table, array_keys(array_slice($params, 0, 2)), array_values($params), 'id = ?')) {
+        $ev = $this->events->trigger($this->getEventName('update'), $params);
+        if (!$ev->getSuccessed() || !$ev->getContinue()) {
+            trigger_error(sprintf("Event userroll\update failed. Returned success = %s, continue = %s", $ev->getSuccessed(), $ev->getContinue()));
+            return false;
+        }
+
+        $params = $ev->getData();
+        if (!is_array($params)) {
+            trigger_error(__METHOD__ . ' save params must be an array!');
+            return false;
+        }
+        unset($params['is_system']);
+
+        $fields = $this->getFieldFromSaveParams($params);
+
+        if ($this->dbcon->update($this->table, $fields, array_values($params), 'id = ?')) {
             $return = true;
         }
 
@@ -247,7 +291,7 @@ implements
         $cn = self::class;
 
         $lt = $this->language->translate($this->leveltitle);
-        
+
         /* @var $copy userRoll */
         $copy = new $cn();
         $copy->setRollName($this->language->translate('GLOBAL_COPY_OF', [$lt], true));
@@ -256,7 +300,7 @@ implements
         if (!$copy->save()) {
             return 0;
         }
-        
+
         $id = $copy->getId();
 
         $permNew = new \fpcm\model\permissions\permissions($id);
@@ -269,4 +313,14 @@ implements
         return $id;
     }
 
+    /**
+     * Returns event base string
+     * @see \fpcm\model\abstracts\dataset::getEventModule
+     * @return string
+     * @since 4.1
+     */
+    protected function getEventModule(): string
+    {
+        return 'userroll';
+    }
 }
